+++ internal/rpc/relation/sync.go
@@ -1,4 +1,4 @@
-package friend
+package relation
 
 import (
 	"context"
